requirements.txt
# requirements.txt
discord.py==2.6.4

motor==3.7.1
pymongo==4.15.5
dnspython==2.8.0

python-dotenv==1.2.1
